fix something and remove tests for now
