+++ README.md
@@ -97,7 +97,7 @@ export default defineConfig({
 If you want to build your own UI instead of using the bundled Vue component, you can consume the shared composable directly:
 
 ```ts
-import { useCopyOrDownloadAsMarkdownButtons } from 'vitepress-plugin-llms'
+import { useCopyOrDownloadAsMarkdownButtons } from 'vitepress-plugin-llms/vitepress-components'
 ```
 
 ---
