+++ src/index.ts
@@ -1,10 +1,2 @@
 export { copyOrDownloadAsMarkdownButtons } from '@/markdown/markdown-it-plugins'
 export { llmstxt as default } from '@/plugin/plugin'
-export {
-	defaultAiProviders,
-	useCopyOrDownloadAsMarkdownButtons,
-} from '@/vitepress-components/use-copy-or-download-as-markdown-buttons'
-export type {
-	MarkdownAiProvider,
-	UseCopyOrDownloadAsMarkdownButtonsOptions,
-} from '@/vitepress-components/use-copy-or-download-as-markdown-buttons'
