+++ tsconfig.bun.json
@@ -7,11 +7,5 @@
 	},
 
 	// Setup linting for tests later
-	"include": [
-		"tests/**/*.ts",
-		"src/vitepress-components/**/*.ts",
-		"lint-staged.config.ts",
-		"bunup.config.ts",
-		"oxlint.config.ts"
-	]
+	"include": ["tests/**/*.ts", "lint-staged.config.ts", "bunup.config.ts", "oxlint.config.ts"]
 }
